+++ app-backend/datamodel/src/main/scala/AnnotationLabel.scala
@@ -38,7 +38,8 @@ final case class AnnotationLabelPropertiesCreate(
 
 @JsonCodec
 final case class AnnotationLabelWithClassesFeatureCollectionCreate(
-    features: Seq[AnnotationLabelWithClasses.GeoJSONFeatureCreate]
+    features: Seq[AnnotationLabelWithClasses.GeoJSONFeatureCreate],
+    nextStatus: Option[TaskStatus] = None
 )
 
 @JsonCodec
@@ -130,7 +131,7 @@ object AnnotationLabelWithClasses {
       id: UUID,
       geometry: Option[Projected[Geometry]],
       properties: AnnotationLabelWithClassesProperties,
-      _type: String = "Feature"
+      @JsonKey("type") _type: String = "Feature"
   ) extends GeoJSONFeature
   @JsonCodec
   final case class GeoJSONFeatureCreate(
@@ -145,12 +146,6 @@ object AnnotationLabelWithClasses {
         properties.description
       )
     }
-  }
-
-  object GeoJSON {
-    implicit val annoLabelWithClassesGeojonEncoder: Encoder[GeoJSON] =
-      Encoder.forProduct4("id", "geometry", "properties", "type")(geojson =>
-        (geojson.id, geojson.geometry, geojson.properties, geojson._type))
   }
 
   /**