+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.64.0] - 2021-05-12
+### Changed
+- Parameterize API Server host in Nginx upstream [#5577](https://github.com/raster-foundry/raster-foundry/pull/5577)
+- Task status updates occur as part of label requests [#5578](https://github.com/raster-foundry/raster-foundry/pull/5578)
+
 ## [1.63.0] - 2021-04-29
 ### Added
 - Added support for Intercom conversations [#5571](https://github.com/raster-foundry/raster-foundry/pull/5571)
@@ -13,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Retried cogification for tifs that fail to become cogs with no appropriate zoom level error [#5573](https://github.com/raster-foundry/raster-foundry/pull/5573)
 - Fixed campaign task list endpoint [#5572](https://github.com/raster-foundry/raster-foundry/pull/5572)
 - Included CORS headers with error responses when appropriate [#5574](https://github.com/raster-foundry/raster-foundry/pull/5574)
+- Task authorization now verifies project relationship [#5576](https://github.com/raster-foundry/raster-foundry/pull/5576)
+
+### Changed
+- Make upload limit check stricter [#5575](https://github.com/raster-foundry/raster-foundry/pull/5575)
 
 ## [1.62.1] - 2021-04-19
 ### Fixed
@@ -928,7 +937,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed issue with clamping imagery whose range was greater than, but included values between 0 and 255 [#4177](https://github.com/raster-foundry/raster-foundry/pull/4177)
 - Included missing `pow` operation for decoding json representations of analyses [#4179](https://github.com/raster-foundry/raster-foundry/pull/4140), [#4155](https://github.com/raster-foundry/raster-foundry/issues/4155)
 
-[Unreleased]: https://github.com/raster-foundry/raster-foundry/compare/v1.63.0...HEAD
+[Unreleased]: https://github.com/raster-foundry/raster-foundry/compare/v1.64.0...HEAD
+[1.64.0]: https://github.com/raster-foundry/raster-foundry/compare/v1.63.0...v1.64.0
 [1.63.0]: https://github.com/raster-foundry/raster-foundry/compare/v1.62.1...v1.63.0
 [1.62.1]: https://github.com/raster-foundry/raster-foundry/compare/v1.62.0...v1.62.1
 [1.62.0]: https://github.com/raster-foundry/raster-foundry/compare/v1.61.2...v1.62.0